+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - (Feature) (Platform) Extend Authorization Policies
 - (Feature) Unify Database structure creation
 - (Feature) (Platform) Sidecar
+- (Feature) (Platform) Sidecar Initial Resources
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService